Ble refactor (#33)
* Now advertising uwb parameters

* Create a function to set the pan ID

* Update API interfaces

* Now updating advertising data for all the uwb settings

* Implement new polling flag

* Add active bit

* Now enabling/disabling uwb through a function

* Add build option to range to neighbors that are inactive

* Fix active uwb flag

* Start creation of custom service

* Created UUIDs

* Defined the header for the Beluga service

* Start defining the beluga service

* Create common api for the beluga service internals

* Define function to encoding uwb configs

* Define function for decoding uwb configs

* Define service

* Define the write function for sync

* Define the service init function

* Add new callback for notif updates

* Now calling the ranging enabled callback

* Write range notify

* Create API for the beluga client

* Write on receive handle

* Write on sent handle

* Write beluga client init function

* Write beluga client sync function

* Write function to discover and assign handles

* Write function for subscribing to ranging

* Marked parameters as unused

* Move update and insertion code into list monitor module

* Start planning out client code

* Create routine for syncing uwb parameters

* Rework the scanning

* Write a routine to sync UWB settings between nodes (TX path)

* Write new AT command for syncing UWB parameters

* Implement the range notification

* Now updating uwb parameters on sync requests

* Update the service to use return values

* Start BLE rewrite

* Update include paths

* Update cmake

* Update makefile

* Moved services and remove old ble app

* Write scan module

* Start advertising module

* Start new app file

* New configs

* Implement client side of the beluga service

* Reimplement server side of the service

* Reimplement the bt stack stuff

* Update the enable/disable functions for Bluetooth

* Implement the ble utility functions

* Fix build problems

* Now checking scanned nodes and advertising nodes independently

* Fix bug that was allowing update_led to fallthrough to the default case

* Update the log levels

* Implement functions to start/stop ble internally

* Update the advertising/scan data

* Add more logging

* Add synchronization and fix ble restart after BLE central connection

* Implement synchronization

* Add logic to disconnect BLE when it is connected to something

* Write docs

* Document the Beluga Service Client API

* Document the BLE service

* Document the advertising part of BLE app

* Document the scan module

* Document ble app

* Document BLE more

* Finish BLE documentation

* Sync now only works if the nodes are on the same PAN network

* Implement GATT for ranging

* Now using double

* Add routines to stop updates to the neighbor list from happening instead of stopping BLE all together

* Update docs

* Clean up merge stuff

+++ Beluga/include/ble/services/beluga_service.h
@@ -132,6 +132,6 @@ int beluga_service_init(struct beluga_service_cb *cb);
  * @return 0 if the data is sent.
  * @return negative error code otherwise.
  */
-int range_notify(struct bt_conn *conn, uint16_t id, float range);
+int range_notify(struct bt_conn *conn, uint16_t id, double range);
 
 #endif // BELUGA_DTS_BELUGA_SERVICE_H
